// autogpt_platform/frontend/src/app/api/__generated__/models/myAgent.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { MyAgentAgentImage } from "./myAgentAgentImage";

export interface MyAgent {
  agent_id: string;
  agent_version: number;
  agent_name: string;
  agent_image?: MyAgentAgentImage;
  description: string;
  last_edited: string;
}